Backend/requirements.txt
# ML & Data Processing
torch>=2.0.0
torch-geometric>=2.4.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
networkx>=3.1

# FastAPI & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Authentication & Security (Clerk JWT verification)
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
httpx>=0.24.0

# Database
supabase>=2.0.0

# File Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
reportlab>=4.0.0
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
